requirements.txt
langchain
langchain-community
langchain-groq
sentence-transformers
chromadb
pymupdf
fastapi
uvicorn
streamlit
scikit-learn
matplotlib
pandas
numpy
jupyter
python-dotenv
Pillow
PyPDF2
langchain-text-splitters
